pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kimera"
version = "1.1.0"
description = "Kubernetes Security Testing Toolkit for Education and Defense"
authors = [{name = "Dynatrace OSS", email = "opensource@dynatrace.com"}]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "kubernetes>=28.1.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "tabulate>=0.9.0",
    "jsonpatch>=1.33",
    "python-decouple>=3.4",
]

[project.urls]
Homepage = "https://github.com/dynatrace-oss/kimera"
Repository = "https://github.com/dynatrace-oss/kimera"

[project.scripts]
kimera = "kimera.exploit_k8s:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "pre-commit>=3.0.0",
    "pip-audit>=2.0.0",
    "safety>=3.0.0",
    "types-pyyaml>=6.0.12",
]

[tool.ruff.format]
docstring-code-format = true


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=kimera --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["kimera"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "kubernetes.*",
    "jsonpatch",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*", "examples.*"]
disallow_untyped_defs = false
warn_unreachable = false

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "S", "D"]
ignore = ["E501", "D100"]  # E501: handled by ruff format; D100: module docstrings not required

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"tests/**" = ["S101", "D"]
"kimera/container/remediations/patcher.py" = ["S603", "S607"]
"kimera/container/core/k8s_client.py" = ["S603"]

[tool.ruff.lint.pydocstyle]
convention = "google"
